Crate tdx_guest

Re-exports

pub use self::tdcall::accept_page;

pub use self::tdcall::get_veinfo;

pub use self::tdcall::TdgVeInfo;

pub use self::tdcall::TdxVirtualExceptionType;

pub use self::tdvmcall::cpuid;

pub use self::tdvmcall::hlt;

pub use self::tdvmcall::print;

pub use self::tdvmcall::rdmsr;

pub use self::tdvmcall::wrmsr;

pub use self::tdvmcall::TDX_LOGGER;

Modules

metadata

tdcall

The TDCALL instruction causes a VM exit to the Intel TDX module.

tdvmcall

The TDVMCALL helps invoke services from the host VMM. From the perspective of the host VMM, the TDVMCALL is a trap-like, VM exit into the host VMM, reported via the SEAMRET instruction flow.

unaccepted_memory

Support for unaccepted memory in TDX guest environments. This module provides mechanisms to manage and accept unaccepted memory regions in TDX guests. The core data structure is EfiUnacceptedMemory, which represents the EFI table header and provides methods to manipulate the unaccepted memory bitmap and perform acceptance operations.

Macros

serial_print

serial_println

Structs

TdAttributes

TdAttributes is defined as a 64b field that specifies various attested guest TD attributes.

Enums

AcceptError

SeptVeError

TopologyError

Traits

TdxTrapFrame

Functions

accept_memory⚠

Accepts a range of physical memory to be used as TDX private memory.

disable_sept_ve

enable_cpu_topology_enumeration

handle_virtual_exception

init_tdx

Performs full initialization of the Intel TDX guest environment.

is_tdx_guest_early

Returns true if the system is identified as an Intel TDX guest during early boot.

reduce_unnecessary_ve

shared_mask

tdx_is_enabled

Type Aliases

TdxGpa