Enum tdx_guest::tdcall::TdCallError

pub enum TdCallError {

    TdxNoValidVeInfo,
    TdxOperandInvalid,
    TdxOperandBusy,
    TdxPageAlreadyAccepted,
    TdxPageSizeMismatch,
    TdxMetadataFieldIdIncorrect,
    TdxMetadataFieldNotWritable,
    TdxMetadataFieldNotReadable,
    TdxMetadataFieldValueNotValid,
    TdxOpStateIncorrect,
    TdxOperandAddrRangeError,
    TdxPageMetadataIncorrect,
    TdxServtdInfoHashMismatch,
    TdxServtdNotBound,
    TdxServtdUuidMismatch,
    TdxTargetUuidMismatch,
    TdxTargetUuidUpdated,
    TdxTdFatal,
    TdxTdKeysNotConfigured,
    TdxTdcsNotAllocated,
    Other,

}

Variants

TdxNoValidVeInfo

There is no valid #VE information.

TdxOperandInvalid

Operand is invalid.

TdxOperandBusy

The operand is busy (e.g., it is locked in Exclusive mode).

TdxPageAlreadyAccepted

Page has already been accepted.

TdxPageSizeMismatch

Requested page size does not match the current GPA mapping size.

TdxMetadataFieldIdIncorrect

The provided FIELD_ID is incorrect.

TdxMetadataFieldNotWritable

Field code and write mask are for a read-only field.

TdxMetadataFieldNotReadable

Field code is for an unreadable field.

TdxMetadataFieldValueNotValid

The provided field value is not valid.

TdxOpStateIncorrect

The TD’s OP_STATE is incorrect for the required operation.

TdxOperandAddrRangeError

Operand address is out of range (e.g., not in a TDMR).

TdxPageMetadataIncorrect

Physical page metadata (in PAMT) are incorrect for the requested operation.

TdxServtdInfoHashMismatch

Service TD hash of TDINFO_STRUCT does not match the currently bound hash.

TdxServtdNotBound

Service TD is not bound.

TdxServtdUuidMismatch

Service TD UUID does not match the currently bound UUID.

TdxTargetUuidMismatch

Target TD UUID does not match the requested TD_UUID.

TdxTargetUuidUpdated

Target TD UUID does not match the requested TD_UUID, but pre-migration target TD UUID does match it.

TdxTdFatal

TD is in a FATAL error state.

TdxTdKeysNotConfigured

TD keys have not been configured on the hardware.

TdxTdcsNotAllocated

TDCS pages have not been allocated.

Other