tdx_guest::tdcall

Struct L2EnterGuestState

pub struct L2EnterGuestState {

    pub rax: u64,
    pub rcx: u64,
    pub rdx: u64,
    pub rbx: u64,
    pub rsp: u64,
    pub rbp: u64,
    pub rsi: u64,
    pub rdi: u64,
    pub r8: u64,
    pub r9: u64,
    pub r10: u64,
    pub r11: u64,
    pub r12: u64,
    pub r13: u64,
    pub r14: u64,
    pub r15: u64,
    pub rflags: u64,
    pub rip: u64,
    pub ssp: u64,
    pub guest_interrupt_status: u16,

}

L2EnterGuestState is used as input and output of enter_l2_vcpu.

It is an array of general-purpose (GPR) register values, organized according to their architectural number, with additional values of RFLAG, RIP and SSP.

Fields
rax: u64rcx: u64rdx: u64rbx: u64rsp: u64rbp: u64rsi: u64rdi: u64r8: u64r9: u64r10: u64r11: u64r12: u64r13: u64r14: u64r15: u64rflags: u64rip: u64ssp: u64guest_interrupt_status: u16