tdx_guest::tdvmcall

Function map_gpa

pub fn map_gpa(gpa: u64, size: u64) -> Result<(), (u64, TdVmcallError)>

MapGPA TDG.VP.VMCALL is used to help request the host VMM to map a GPA range as private or shared-memory mappings. This API may also be used to convert page mappings from private to shared. The GPA range passed in this operation can indicate if the mapping is requested for a shared or private memory via the GPA.Shared bit in the start address.