tdx_guest::tdvmcall

Function setup_event_notify_interrupt

pub fn setup_event_notify_interrupt(
    interrupt_vector: u64,
) -> Result<(), TdVmcallError>

The guest TD may request that the host VMM specify which interrupt vector to use as an event-notify vector.

This is designed as an untrusted operation; thus, the TD OS should be designed not to use the event notification for trusted operations.

Example of an operation that can use the event notify is the host VMM signaling a device removal to the TD, in response to which a TD may unload a device driver.

The host VMM should use SEAMCALL [TDWRVPS] leaf to inject an interrupt at the requestedinterrupt vector into the TD VCPU that executed TDG.VP.VMCALL via the posted-interrupt descriptor.